look for saves when uninstalling

+++ uninstall.sh
@@ -11,6 +11,17 @@ fi
 
 mv cloudpie/roms ~/
 
+echo "looking for saves"
+if pgrep sync.sh || pgrep rclone; then
+    pkill sync.sh
+    pkill rclone
+fi
+
+if [ -e ~/cloudpie/save/cloud.txt ]; then
+    echo "unmounting saves failed"
+    exit 1
+fi
+
 rm -rf cloudpie
 rm -rf retroarch
 rm -rf .config/retroarch
